+++ packages/connect-miniprogram/src/connect/protocol/normalize.ts
